+++ packages/mantine/src/MantineValueEditor.tsx
@@ -22,7 +22,6 @@ export const MantineValueEditor = (allProps: MantineValueEditorProps): React.JSX
     title,
     className,
     type,
-    inputType,
     values = [],
     listsAsArrays,
     separator,
@@ -32,18 +31,19 @@ export const MantineValueEditor = (allProps: MantineValueEditorProps): React.JSX
     selectorComponent: SelectorComponent = allProps.schema.controls.valueSelector,
     validation: _validation,
     extraProps,
+    inputType: _inputType,
     parseNumbers: _parseNumbers,
     ...propsForValueSelector
   } = allProps;
 
-  const { valueAsArray, multiValueHandler, valueListItemClassName } = useValueEditor(allProps);
+  const { valueAsArray, multiValueHandler, valueListItemClassName, inputTypeCoerced } =
+    useValueEditor(allProps);
 
   if (operator === 'null' || operator === 'notNull') {
     return null;
   }
 
   const placeHolderText = fieldData?.placeholder ?? '';
-  const inputTypeCoerced = ['in', 'notIn'].includes(operator) ? 'text' : inputType || 'text';
 
   if (
     (operator === 'between' || operator === 'notBetween') &&
@@ -81,6 +81,7 @@ export const MantineValueEditor = (allProps: MantineValueEditorProps): React.JSX
                 i
               )
             }
+            popoverProps={{ withinPortal: false }}
             {...extraProps}
           />
         );
@@ -201,6 +202,7 @@ export const MantineValueEditor = (allProps: MantineValueEditorProps): React.JSX
             const dateArray = dates.map(d => (d ? dayjs(d).format(dateFormat) : ''));
             handleOnChange(listsAsArrays ? dateArray : dateArray.join(','));
           }}
+          popoverProps={{ withinPortal: false }}
           {...extraProps}
         />
       );
@@ -217,6 +219,7 @@ export const MantineValueEditor = (allProps: MantineValueEditorProps): React.JSX
           onChange={d =>
             handleOnChange(d ? dayjs(d).format(dateTimeLocalFormat) : /* istanbul ignore next */ '')
           }
+          popoverProps={{ withinPortal: false }}
           {...extraProps}
         />
       );
@@ -232,6 +235,7 @@ export const MantineValueEditor = (allProps: MantineValueEditorProps): React.JSX
         onChange={d =>
           handleOnChange(d ? dayjs(d).format(dateFormat) : /* istanbul ignore next */ '')
         }
+        popoverProps={{ withinPortal: false }}
         {...extraProps}
       />
     );