+++ packages/react-querybuilder/src/utils/parseSpEL/index.ts
@@ -1,1 +1,8 @@
+/**
+ * Converts a SpEL string expression into a {@link index!DefaultRuleGroupType DefaultRuleGroupType}
+ * or {@link index!DefaultRuleGroupTypeIC DefaultRuleGroupTypeIC}.
+ *
+ * @module parseSpEL
+ */
+
 export * from './parseSpEL';